feat: automatically handle dates in query string
When a date is passed to the query string, especially when using the _useDateType_ option, it silently fails to convert the date.
Since the URL and query is not passed to the interceptor, there is no opportunity to correct this.
This will automatically convert dates to an ISO string which should be accepted by most platforms.