chore: specify pnpm version in package.json
Problem
=

It's possible to use a different version of pnpm than the one that is
used in CI. This can lead to inconsistencies between the local
development environment and the CI environment.

Solution
=

Specify the version of pnpm that is used in the `engines` field of
`package.json`.

This results in an error message similar to the following when a
different version of pnpm is used:

```sh
$ corepack prepare pnpm@latest --activate
$ pnpm i
 ERR_PNPM_UNSUPPORTED_ENGINE  Unsupported environment (bad pnpm and/or Node.js version)

Your pnpm version is incompatible with "/Users/mikavilpas/git/openapi-ts".

Expected version: 8.15.7
Got: 9.9.0

This is happening because the package's manifest has an engines.pnpm field specified.
To fix this issue, install the required pnpm version globally.

To install the latest version of pnpm, run "pnpm i -g pnpm".
To check your pnpm version, run "pnpm -v".

$ corepack prepare pnpm@8.15.7 --activate
```