refactor: use `@ts-expect-error` to keep error suppression up to date
Problem
=

Using `@ts-ignore` will suppress a type error, but if for some reason
the type error no longer occurs in the future, it's difficult to notice
this. Typically the `@ts-ignore` comment is left in the codebase, which
can weaken the level of type checking in some cases.

Solution
=

Use `@ts-expect-error` instead. It does the same thing but also
complains if it is being used to suppress an error that does not exist.

You can read more about it here

- https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-9.html#-ts-expect-error-comments
- https://dev.to/maafaishal/ts-expect-error-over-ts-ignore-in-typescript-5140