refactor: eslint uses `import type` where possible
This commit adds an eslint rule
`@typescript-eslint/consistent-type-imports` that warns (and
automatically fixes) when `import` is used instead of `import type` for
importing types.

This has minor performance benefits at build time and maybe at runtime,
but can also communicate the intent of the import more clearly.

For a nice developer experience, it's recommended to enable autofixing
fixable eslint issues so that the developer never has to worry about
this - it's done completely automatically when the file is saved.
Autoformatting with prettier is also recommended.

- https://typescript-eslint.io/blog/consistent-type-imports-and-exports-why-and-how/
- https://typescript-eslint.io/rules/consistent-type-imports/