response type void when no content in schema.
After this change response type is set to void when no content schema is provided, regardless of status code of response.

Before this change, it was only set to void for status code 204.

The spec (https://swagger.io/docs/specification/v3_0/describing-responses/) does not say that an empty response body is is only valid for 204 responses, it is just used as an example.