fix!: rearrange order of parameters
This is a somewhat breaking change. Previously, the deprecated and description field have been overwritten by the schema property of parameters. This leads to parameters that were marked as deprecated on the parameter-level to not be marked as deprecated in the generated output IF the schema was not marked as deprecated. Local configuration should overwrite schema configuration. This is a breaking change, as it introduces different (valid) documentation compared to how it was handled previously. The general runtime of outputs should not be affected.